Changed variable names, cleaned up code
Changed a few variable names to make them more descriptive and clear. Also reorganized a few bits of code so everything is in the same place and is a bit better organized.

I'm submitting this as a new branch so the code can be reviewed before committing to the changes. This is a good way to propose changes without making them permanent.