Added basic hanging control
I added a hangingControl function to teleop, which implements hanging. It also has feed-forward functionality built into it, currently controlled with a button in teleopPeriodic. Hitting X toggles feed-forward on or off. Variables are declared and then values are assigned to them in RobotInit to make this work. If we decide we always want this feature activated (or never want it activated), we can simply set it to false.