Removed 2016 refs, merged intake/feeder
Removed the "is2016Robot" variable and all conditionals regarding it as the robot will be taken apart soon. Fixed a bug with the intake and feeder code which prevented each other from running - now combined into single class. Deadband protection added for that function due to problems with analog sticks on shitty Xbox controllers. PID needs fixing, working on that