PID Fixes
The PID function was not calling the right methods to assign the constants. The speed setpoint was not properly calculated per the Talon instructions. The encoder CPR was not accounted for. All should be fixed with this update. Also added a Boolean so we can see on the screen if the shooter is supposed to be on.