PID Constant Tweaking, Competition Changes
Fixed Test Mode so it will work to test shooter motors without having to
edit code temporarily. Various PID constants were tested and changed
last night, now included in comments.

The competition robot was tested and one hardware change was found. A
conditional, isPracticeRobot, is used to configure the robot differently
if it is or is not the practice robot. Jake is working on a better way
to do this, but for now the constant is changed to "true" or "false" at
compile time.