Autonomous Mode!!!
Center Gear is working, Right Gear is about 80% there. Left Gear should
be basically working as soon as Right Gear is.

The switch case structure is the proper way to write autonomous
sequences. The use of "if you aren't there, go" control instead of P
loops for distance is bad, but it's all we have time for before
Waterbury.