PID Tweaks, Gyro + Auto Code
Changed PID values to last experimental working values. Will need to tweak throughout competition. We need to look into a method of *saving* values to the rio between matches.

Gyro Calibrate defaults to on, so that we will never forget to do it.

The "default auto" now drives forward for a fixed distance. Gyro Test Auto can be used to make an auto that does nothing. The default auto is completely untested but an exact copy of gyro test auto code.

I added in some error handling code for the encoder but I commented it out because it hasn't been tested. 

I did not add rate limiting code.